♻️ refactor(User): User 엔티티 totalCoin type int로 변경 (#4)

+++ src/main/java/com/sparta/devquiz/domain/user/entity/User.java
@@ -51,7 +51,7 @@ public class User extends BaseTimeEntity {
     private UserRole role;
 
     @Column(nullable = false)
-    private long totalCoin;
+    private int totalCoin;
 
     @Column
     private LocalDateTime deletedAt;
